requirements.txt
pyyaml
lxml
bz2file
python-dateutil
psycopg2